feat(components): add props to Folder and File components
Updated the Folder component to accept `folderName` and `folderColor` properties and included currentColor feature considering these properties. Also added the ability to toggle open/close the folder with the `isOpen` computed property.

Created a new File component that accepts a `to` property for routing purposes. This component was also added to the HobbiesView.

Adjusted the HobbiesView to accommodate these changes.